Start of Version 0.2.3
Made the files for Version 0.2.3
Checking the maps and lore there it would be possible for Xerxes to collect the Makkabium and Bala-Axerium tablet pieces in any order. So the story is now structured to allow the player to also to collect the Makkabium and Bala-Axerium tablet pieces.

Most of the dialog is done - still need to do some Makkabium, Yusinzia Rebellion and Xartabana dialog. While I won't do all of the Anti-Stealth Tablet version of the events now. I will make the base skeleton and dialog for when I need to implement them in version 0.3.2.
Xartabana will act as the hub for the player party in the Eastern Jamesos Realm as it is the only place that wouldn't attack the protagonists outright.